third_party/WebKit/LayoutTests/fast/block/float/override-property-float.html

<!doctype html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<style>
  #test {
      width:100px;
      height:100px;
      background:green;
      float:right;
      float:center;
  }
</style>
</head>
<body>
    <div id="test">TEXT</div>
</body>
<script>
description('If this test passes, the green box should be on the right side of page.');
var element = document.getElementById('test');
shouldBeEqualToString('window.getComputedStyle(element, null).getPropertyValue("float")', 'right');
</script>
</html>
