third_party/WebKit/LayoutTests/fast/block/float/override-property-float-expected.txt

If this test passes, the green box should be on the right side of page.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.getComputedStyle(element, null).getPropertyValue("float") is "right"
PASS successfullyParsed is true

TEST COMPLETE

TEXT
